worker/requirements.txt
yfinance>=0.2.30
pandas>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
